+++ ChangeLog.md
@@ -27,7 +27,7 @@ and simplify code. See the Venice *JMH* benchmark *ReflectionBenchmark.java*
 
 ### Added
 
-- ...
+- pre-loaded modules to honor the 'macroexpandOnLoad' flag (Java integration API)